minimum performance upgrades
