+++ lib/widgets/profile/editprofile_inputs.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-
 import 'package:get/get.dart';
 import 'package:wisdrive/constraints/helper_functions.dart';
 import 'package:wisdrive/constraints/popup_messages.dart';
